+++ src/copytuner-bar.ts
